fix: better side nav
