+++ .eleventy.js
@@ -11,21 +11,7 @@ const pluginSyntaxHighlight = require("@11ty/eleventy-plugin-syntaxhighlight");
 const { EleventyRenderPlugin } = require("@11ty/eleventy");
 const pluginNavigation = require("@11ty/eleventy-navigation");
 const PostCSSPlugin = require("eleventy-plugin-postcss");
-
-  async function copyRecursive(from, to) {
-    const { copyFile, lstat, mkdir, readdir } = fs.promises;
-    await mkdir(to, {recursive: true});
-    for (const element of await readdir(from)) {
-      const _from = path.join(from, element);
-      const _to = path.join(to, element);
-      const stat = await lstat(_from);
-      if (stat.isFile()) {
-        await copyFile(_from, _to);
-      } else {
-        await copyRecursive(_from, _to);
-      }
-    }
-  }
+const RHDSPlugin = require("./_plugins/rhds.cjs");
 
 module.exports = function(eleventyConfig) {
   // Copy the `assets` folder to the output
@@ -37,6 +23,8 @@ module.exports = function(eleventyConfig) {
   eleventyConfig.addPlugin(pluginNavigation);
   eleventyConfig.addPlugin(EleventyRenderPlugin);
   eleventyConfig.addPlugin(PostCSSPlugin);
+  eleventyConfig.addPlugin(RHDSPlugin);
+
   eleventyConfig.addDataExtension("yaml", contents => yaml.load(contents));
 
   eleventyConfig.addFilter("readableDate", dateObj =>
@@ -95,20 +83,6 @@ module.exports = function(eleventyConfig) {
     slugify: eleventyConfig.getFilter("slugify")
   });
   eleventyConfig.setLibrary("md", markdownLibrary);
-
-  eleventyConfig.on('eleventy.before', async () => {
-    const { copyFile } = fs.promises;
-    console.log('Copying base RHDS styles...');
-    const globalStylesIn = path.join(require.resolve('@rhds/tokens'), '..','..','css', 'global.css');
-    const globalStylesOut = path.join(__dirname, 'assets', 'css', 'rhds.css');
-    if (!fs.existsSync(globalStylesOut))
-      await copyFile( globalStylesIn, globalStylesOut);
-    console.log('Copying RHDS elements assets...');
-    const from = path.join(require.resolve('@rhds/elements'), '..');
-    const to = path.join(__dirname, 'assets', '@rhds', 'elements');
-    await copyRecursive(from, to);
-    console.log('  ...done')
-  })
 
   // Override Browsersync defaults (used only with --serve)
   eleventyConfig.setBrowserSyncConfig({
